+++ CHANGELOG.md
@@ -14,12 +14,14 @@
 + [x] Calendar: add class `day-border` and attribute `data-day-border="true|false"`
 + [x] Validator: fix error if value is undefined  
 + [x] Validator: function `date` now support additional input attribute `data-value-format` for non ECMAScript dates
++ [x] Validator: function `date` now support additional input attribute `data-value-locale` for non ECMAScript dates
 + [x] Z-index: set equal z-index for `appbar`, `bottomnav`, `bottomsheet`, `tabsmaterial`
 + [x] Carousel: rename attribute `data-bullet-style` to `data-bullets-style`
 + [x] Carousel: add attribute `data-bullet-size` with values `default`, `mini`, `small`, `large`
 + [x] Carousel: add style `cycle`
 + [x] Popover: set default value for `data-popover-timeout` to `10`, issue #1277
 + [x] Storage: Objects `storage` and `session storage` combined into one object. Access to objects remained unchanged: `Metro.storage`, `Metro.session`
++ [x] String: `String.toDate` now support `locale` as second parameter: `"21 грудня 1972".toDate('%d %m %y', 'uk-UA')`
 
 ### 4.2.30
 + [x] Toast: add function `init(options)`, now you can set toast `top` position and `distance`. See docs for details.